fix(enricher): cast double for pgvector cosine similarity result
PostgreSQL <=> operator returns double precision. SimilarTicketRow
and SimilarTicket were typed float, causing Npgsql InvalidCastException
on every FindSimilarAsync call that returned rows.

+++ src/ApartmentTriage.Application/Agents/Enricher/SimilarTicket.cs
@@ -5,6 +5,6 @@ namespace ApartmentTriage.Application.Agents.Enricher;
 public sealed record SimilarTicket(
     Guid           TicketId,
     TicketCategory Category,
-    float          CosineSimilarity,
+    double         CosineSimilarity,
     DateTime       CreatedAt
 );
